requirements.txt
django
djangorestframework
requests
black
flake8
python-dotenv
swagger_auto_schema